guanfu_backend/pyproject.toml
[project]
name = "guanfu-backend"
version = "0.1.0"
description = ""
requires-python = ">=3.11"
dependencies = [
  "fastapi (>=0.118.0,<0.119.0)",
  "uvicorn (>=0.37.0,<0.38.0)",
  "python-dotenv (>=1.1.1,<2.0.0)",
  "sqlalchemy (>=2.0.43,<3.0.0)",
  "pydantic-settings (>=2.11.0,<3.0.0)",
  "psycopg2-binary (>=2.9.10,<3.0.0)",
  "alembic>=1.16.5,<2.0.0",
  "redis>=5.0.0,<6.0.0",
  "asyncio-mqtt>=0.16.0,<1.0.0",
]
